Hack up a link to the suites directory
The generated makefile compiles the unit tests from the root
directory. This has the advantage of not requiring different
paths (-I, -L). For the unit tests, unfortunately, this results in
incorrect debugging information for the common .function files that
get pulled in via the test generator rather than an #include
directive. For these files, the .c file to compile contains #line
directives with a path that's relative to the tests directory, and
therefore the debugging data lists them as e.g.
`.../build/suites/main_test.function` rather than
`.../source/tests/suites/main_test.function`.

As a workaround, create a symbolic link .../build/suites that points
to the correct directory. In fact, this is only the correct directory
if the test doesn't come from a submodule, so this is only a partial
workaround.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>