Rule change: loosen the rule for static functions
The rule that static functions must start with the module name is not
currently systematically followed: some start with `mbedtls_` and some don't
include the module name. Allow both deviations.

Rationale: the rule is already widely ignored. Using the `mbedtls_` prefix
on static functions makes it easier to export the function later, which is
necessary to write unit tests. Not using a prefix is convenient for
small auxiliary functions that are very specific to one module.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>