Require help with --help, not with no options
Make a stronger requirement that --help should print usage information: it's
more convenient than a comment. (But don't make it an absolute requirement
in case it's too annoying to support long options.)

For many scripts, running with no options is the normal thing, so don't
suggest printing help in that case.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>