Improve discussion of Python versions
Move the discussion about Python to the Python-specific section.

In the LTS, only user-facing scripts need to remain backward compatible with
old versions. In 2.28, that's just config.py, so discuss this case
specifically. (We may want to move this information to config.py itself, but
for now record it here.)

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>