First stab at a test suite for mbedtls-rewrite-branch-style
The test code is hastily written and may not quite work yet on a clean
setup.

The method structure and naming are a bit haphazard and could use some
cleaning up and uniformizing.

The test data comes from a mixture of reference results that must have been
manually validated and pushed to the repository, and hard-coded commit sha
values. It should probably be just one or the other (but with instructions
on how to prepare the reference results).

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>