Ensure new commits have the desired content after cherry-pick
Conflict resolution during cherry-pick is not guaranteed to give the result
we want, which is that files modified on both sides (restyled in the branch
we're building, modified in the next commit) have the content from the new
commit except for styling. It seemed to work for me with git 2.25.1 but
failed with git 2.38.1 (and possibly different diff/merge strategy
configurations). For robustness, force changed files to their content
in the next commit.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>