Use branches for test data and archive the target branch used
Don't hard-code any commit ID in the script, but instead rely on branch
names in the upstream repository. Reference outputs need to be
archived (since each run produces different commit dates and thus has
different shas), and we might as well use the same mechanism for reference
inputs for simplicity.

When we test that the output matches a reference output, we need to use
repeatable inputs. So instead of relying on
features/new-code-style/development and such (which is updated from time to
time), archive the specific version of the target branch used for the
reference output.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>