Deal with deleted files in restyling script
When restyling a branch, files that are deleted in the branch cause
conflicts with the restyling to them that has taken place in the rebase.
These conflicts are not solved by the -Xtheirs option.

Fix this by getting a list of all files that are deleted by the branch
but also restyled in the target and resolve the conflicts with git rm in
the script.

Signed-off-by: David Horstmann <david.horstmann@arm.com>