Fix handling of commits that delete a C file
Commits that delete a styled file would cause a modify/delete conflict
during cherry-picking. This is a "high-level" conflict which cannot be
resolved automatically by the recursive strategy (and defining a custom
strategy that does what we want is hard and undocumented).

To avoid such a conflict, apply the cherry-pick after locally deleting the
offending files.

Thanks for David Horstmann for identifying the problem and exploring the
solution space.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>