Guidance on contribution content and structure
Give some guidance on where to contribute, what content is expected, and how
content should be structured.

This complements the existing contribution guidance: CONTRIBUTING.md in the
mbedtls repository, and the guide to the review process from a contributor's
point of view. It might make sense to combine this information in a single
document, but that will be for later.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>