Deal with no-name-change edge case
When files are moved to the exact same location in the framework repo
as they were in Mbed TLS, there is a problem.

Since no git change has taken place on the Mbed TLS side, the most
recent change to the file is its deletion when a previous set of files
were moved to the framework (since all other files are deleted in this
kind of move).

As a result the moved files are deleted and do not appear in the final
branch. Deal with this edge case explicitly by taking files which are
not renamed from the file map and checking them out to the state in
the temporary branch, before amending the resulting commit.

Signed-off-by: David Horstmann <david.horstmann@arm.com>