Fix ambiguous "generated by" line and add "prepare" target
Add shell quoting to the "generated by" line so that it can be copy-pasted
into a shell, even if some arguments contain spaces and other special
characters.

Add a new target `prepare`, alias `dep`: `make prepare` or `make dep`
regenerates the makefile. This might not yet work in all arrangements of the
build directory relative to the source directory; it works at least when
`mbedtls-prepare-build` is invoked inside the source directory.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>