Fix bug in mbedtls-kick-ci
When a number is supplied with no prefix, we are meant to assume Mbed
TLS by default. This we do, but we also remove the first digit of the PR
number inadvertantly.

Fix this so that the PR number is interpreted correctly as referring to
Mbed TLS and starts the correct CI.

Signed-off-by: David Horstmann <david.horstmann@arm.com>