Add the new openci, keeping the old one as "linaro"
We are moving to a new instance of OpenCI. Keep the name `openci` on the
command line to refer to the preferred instance, which is changing from the
old one maintained by Linaro to the new one maintained by Arm (both on
behalf of TrustedFirmware). Introduce the name `linaro` to refer to the old
one.

Keep the name of the variables in the config file, because existing configs
with a token for the old instance should keep working. Users should add
credentials for the new CI as `JENKINS_CLOUDBEES_xxx`.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>